Fix sshb_projects UPSERT failures for large/decimal money fields
- Widen jdacost and connections from smallint to bigint (smallint
  capped at 32,767, but JDA Cost is a Naira amount and connections
  is a household count, both of which routinely exceed that). Adds
  migrations/widen_sshb_projects_smallints.sql to apply the change
  to live DBs; updates supabase-setup-img.sql for fresh installs.

- Round numeric form inputs bound for integer columns. Form fields
  use Number(e.target.value), which yields floats like 5000000.77;
  Postgres rejects those for bigint/integer with 22P02. pickColumns
  now consults a per-table intColumns list and Math.rounds matching
  values at the UPSERT boundary, so calculations and future inputs
  are also protected.

+++ supabase-setup-img.sql
@@ -23,7 +23,7 @@ create table public.sshb_projects (
   rag                text,
   size               numeric,
   updated_at         timestamptz not null default now(),
-  connections        smallint,
+  connections        bigint,
   loi                boolean default false,
   jda                boolean default false,
   credit             boolean default false,
@@ -40,7 +40,7 @@ create table public.sshb_projects (
   "evidenceCompliance" integer default 100,
   "pvCapacity"       numeric default 0,
   duration           integer default 0,
-  jdacost            smallint
+  jdacost            bigint
 );
 
 create table public.sshb_team_members (
